Sort trips by departure/arrival dates
Store data shaping vs data displaying helper functions in different
files.
