+++ src/displayData.js
@@ -1,0 +1,5 @@
+import moment from 'moment';
+
+export const displayISOTime = (time) => {
+  return moment(time).format('dddd, MMMM Do YYYY, HH:mm');
+};
